Move prompt template tests to androidUnitTest
- commonTest runs as plain JVM (no Context), so these are moved to androidUnitTest and run under Robolectric.
- Update `CreateShareablePromptUseCase` to depend on `ComposeResourcePromptTemplateRepository`.
- Move prompt strategy and use case tests from `commonTest` to `androidUnitTest` to support Robolectric and Android resources.
- Add `initComposeResourcesContext` utility to initialize `AndroidContextProvider` for unit tests.
- Enable `isIncludeAndroidResources` in `build.gradle.kts` for the shared module.
- Refactor prompt creation tests to use `runTest` and fetch templates from the repository.
