+++ shared/build.gradle.kts
@@ -116,6 +116,11 @@ android {
     buildFeatures {
         buildConfig = true
     }
+    testOptions {
+        unitTests {
+            isIncludeAndroidResources = true
+        }
+    }
 }
 
 mavenPublishing {
